Fix for serialEvent example
In the example is stated that the function is run between one loop and the next, but actually the call to the function was missing.
The comment also state that the response can be delayed using a delay in the loop,so I think that the way it should be is so by only
adding a call to the function serialEvent as first operation in the loop. I so added this call.