Add GitHub Actions workflow to check for commonly misspelled words
This workflow uses codespell to check all files in the repository for commonly misspelled words. The "blocklist" strategy used by codespell leads to a minimal number of false positives at the cost of an increased likelihood of false negatives when compared to the more common "smart allowlist" strategy used by spell checkers.

In the event of a false positive, add the problematic word, in all lower case (regardless of the case of the occurence of the word in the repository) to extras/codespell-ignore-words-list.txt.

The workflow is triggered:
- On pull requests
- On push (to validate commits in preparation for a PR)
- On weekly schedule (to check for new misspelled word detections as a result of updates to the codespell dictionary)