Remove compilations of sketches using pulseIn() for Portenta from CI workflow
The pulseIn() function has not yet been implemented for the Portenta H7 boards in the core library of their "Arduino Mbed OS Boards (nRF52840 / STM32H747)" platform.

Once the pulseIn() function has been implemented for Portenta H7 and a release of the platform made containing that implementation, this commit should be reverted.