Update all URLs in example comments
Some of the URLs were broken.

Almost all of the URLs had changed, but were still working due to redirects. However, these redirects might stop working
at some time in the future (as happened during the last round of website reorganization) so it's safest to point them to
the final location.