Use modern action version in sketch compilation workflow
The GitHub Actions action used to do a "smoke test" compilation of the examples has graduated from an "experimental"
project to a stable state and been moved from its original home in the `arduino/actions` repository to a dedicated
permanent home at `arduino/compile-sketches`. The copy in the previous repository is unmaintained and deprecated and its
use results in a warning of such in the workflow run summary and logs.