Use modern `arduino/compile-sketches` API
The original API of the `arduino/compile-sketches` GitHub Actions action used by the "smoke test" sketch compilation workflow was based on Bash array-style space separated lists. This could not to advanced use cases and so a new approach of defining the data structures in YAML documents and then passing those to the action inputs was settled on. At the time the workflow was written, this had not yet been implemented for the `sketch-paths` input.

The old API is now deprecated and so the workflow is here updated to the new API.