Use modern board-specific Mbed OS Boards platforms
Since the time the workflow was written, the "Arduino Mbed OS Boards" platform was split into a variant for each of the
boards "families" supported by that platform. The benefit of this split is a reduction of installation and compilation
times. The previous `arduino:mbed` platform is still available to ease the transition to the new platforms, but it is
deprecated.