Enable `pulseIn()` sketches for Portenta in sketch compilation workflow
At the time the "Compile Examples" CI workflow was written, the Portenta H7 board did not have a `pulseIn()` function.
Although this was a legitimate bug, since this was an issue with the `arduino/ArduinoCore-mbed` code base, already
tracked there, and unrelated to the code hosted in this repository, it was determined that the best approach would be to
configure the workflow to not compile the examples that used `pulseIn()` for the Portenta H7.

Since that time, `pulseIn()` has been implemented for Portenta H7 and so we can now compile these sketches for that board.