Remove "pulsein" board attribute from "Compile Examples" CI workflow
This attribute was used to configure the workflow to skip compilation of examples that used `pulseIn()` for boards that
were already known to not provide this functionality.

The only such board was the Portenta H7 and that has since gained a `pulseIn()`. Since all boards now have `pulseIn()`,
the attribute no longer serves a purpose and can be removed from the workflow.