Configure "Compile Examples" workflow to skip incompatible sketch/board combo
Pin A5 is used by the "StringComparisonOperators" example. The Nano RP2040 Connect uses the u-blox NINA-W102's ADC for
pins A4-A7. This means those pins are only usable in a sketch when it contains an `#include` directive for the WiFiNINA
library. This results in compilation of that example failing for the Nano RP 2040 Connect. The need for this `#include`
directive is surely an intentional design decision, and thus not likely to change.

Although it is possible to add compatibility to the sketch with a preprocessor conditional, the added complexity in what
is intended to be a very beginner friendly example is likely not worth it. Another option would be to change the sketch
to use a different pin, but that sort of change does not fall within my remit.

The error message the user will experience if they attempt to compile this sketch for their board is reasonably good so I
suppose the realistic best way forward is to consider this not a bug and just skip the sketch.