Break code formatting command into multiple lines
A script included in the repository formats the code in the examples according to Arduino's standard style.

Since the formatter tool does not have a recursive capability, `find` must be used. This results in a fairly long and
complex command.

Adding line breaks to split the command into its individual arguments, with indentation to indicate the structure of
these arguments makes the command easier to understand.