Use globbing for extension mapping in formatting script
The repository includes a script that formats the example sketch code. This uses a `find` command to recurse through the
example folders and run the formatter on each file with a relevant extension.

Previously a regular expression was used for that matching. Although perfectly functional, the same can be achieved in
an easier to understand and maintain manner using globs.