Make formatter script check availability of tool
The repository includes a script that formats the code in the examples.

When running this script locally, the contributor must have the formatter tool installed and in the system PATH.
Previously this dependency was only explained by a comment in the script, which is not likely to be seen by the
contributor in advance of running it. This means they would encounter a somewhat cryptic error message.

A more contributor friendly approach is for the script to check for the availability of the tool, then exit with a
friendly error message if it is not found.