Use formatter script in CI workflow
The repository includes a script that formats the code in the example sketches.

Previously, this script was only provided to make it easy for contributors to achieve compliant code formatting while
preparing a pull request.

In addition to this effort to facilitate compliance, enforcement is implemented by a GitHub Actions workflow.

Since the formatter tool does not have the directory tree recursion capability needed to process all the files in the
repository, it is necessary to implement such capability using a `find` command. Previously this was done both in the
script and the workflow.

`find` is ridiculously complex and unintuitive, prone to confusing bugs. For this reason, it will be best to avoid
maintaining redundant code for recursing through the example code files. This is accomplished by simply running the
script from the workflow. Even though the two have different goals (making formatting compliant vs checking whether
formatting is compliant), the two are easily aligned by formatting the files in the workflow, then checking for a diff.
This approach has been widely used in other Arduino Tooling Team formatting check workflows and I find the diff is very
effective at communicating which changes are required, likely more so than the approach previously used in the workflow.