Use ClangFormat as code formatter for examples
The code formatting tool used by the Arduino IDE 1.x "Auto Format" feature is "Artistic Style". This was changed to
ClangFormat in Arduino IDE 2.x. Now that the stable Arduino IDE 2.0.0 release has been made, this is now the official
code formatting tool for use with Arduino sketches.

Due to differences in the two tools, some unavoidable minor changes occurred in the formatting style. This meant that
the Arduino IDE 2.x user may experience some unexpected changes to the unmodified official example sketches when running
an "Auto Format". So the examples in this repository must be formatted to comply with the updated Arduino code style.

The changes are mostly the result of ClangFormat being more "opinionated" than "Artistic Style" was under the fairly
light-handed formatter configuration used by Arduino IDE 1.x Auto Format. Due to this factor, formatting changes
required to bring the examples into compliance with the new style mostly do not conflict with the Arduino IDE 1.x style.
The only exception is the more complex code of the ArduinoISP sketch. In this case, the sketch is more of a utility
program than an example, so the formatting changes will not be impactful for users who continue to use the outdated
Arduino IDE version.