Add new boards to sketch compilation workflow
The repository infrastructure includes a GitHub Actions workflow that compiles the examples for significant official
Arduino boards. This provides a automated "smoke test" to easily get some basic validation of the sketches.

Since the time the workflow was created, several new official boards have been released. Some of these boards are
significantly different from the boards the workflow was compiling the examples for, which meant the workflow was not
providing adequate coverage for the new boards.

The workflow is hereby configured to also compile the examples for the significant new boards.