Document incomplete nature of "p08_DigitalHourglass"
The sketches under the "10.StarterKit_BasicKit" folder are provided to accompany the Arduino Projects Book of the
Arduino Starter Kit. The Arduino Projects Book explains the code of these sketches and gives instructions for their
usage.

In addition to owners of the Arduino Starter Kit, other users might use these prominent sketches as learning references.
Those users won't have the context that would be provided by the Arduino Projects Book. For this reason, explanatory
comments have been added to these sketches.

The "p08_DigitalHourglass" is unique in that it is intentionally left incomplete. The Arduino Projects Book instructs
the user to add their own code to indicate the completion of the timer. It is important that this code be added, since
the incomplete code does not halt the LED control code at the end of the hour. If the incomplete code is used unchanged,
the LED control code will call `digitalWrite` on additional pins.

Previously, the comments in the sketch did not explain the incomplete nature of the sketch. Comments are hereby added to
explain this to users who don't have the context provided by the Arduino Projects Book.